Remove size from Store.pagination
Decided it's a bit confusing to keep track of
properties that are defined in different parts
of the store. For now, decided to keep track
of size in search because we first use the size
in the <SearchForm /> to construct the initial
frontendURL that takes the user to the
/search page and then kicks off the Backend
API Search flow.